Extract SOAP stubs to separate project
