+++ api-stubs/api-stubs-soap/src/main/resources/blogPostService.wsdl
