+++ pom.xml
@@ -5,9 +5,10 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>de.samples.api-comparison</groupId>
-  <artifactId>agregator</artifactId>
+  <artifactId>aggregator</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>pom</packaging>
+  <name>API Comparison Samples</name>
 
   <properties>
     <maven.compiler.source>21</maven.compiler.source>
