--- service-consumer-spring/src/main/java/de/samples/apicomparison/consumer/clients/soap/stub/messages/package-info.java
@@ -1,2 +1,0 @@
-@jakarta.xml.bind.annotation.XmlSchema(namespace = "http://samples.de/spring/soap/blog/messages")
-package de.samples.apicomparison.consumer.clients.soap.stub.messages;
