--- service-consumer-spring/src/main/java/de/samples/apicomparison/consumer/clients/soap/stub/model/BlogPost.java
@@ -1,219 +1,0 @@
-
-package de.samples.apicomparison.consumer.clients.soap.stub.model;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.xml.datatype.XMLGregorianCalendar;
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlID;
-import jakarta.xml.bind.annotation.XmlSchemaType;
-import jakarta.xml.bind.annotation.XmlType;
-import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-
-
-/**
- * <p>Java-Klasse für BlogPost complex type.
- * 
- * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
- * 
- * <pre>{@code
- * <complexType name="BlogPost">
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="id" type="{http://samples.de/spring/soap/blog/types}UUID"/>
- *         <element name="title" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         <element name="content" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         <element name="timestamp" type="{http://www.w3.org/2001/XMLSchema}dateTime"/>
- *         <element name="author" type="{http://samples.de/spring/soap/blog/types}Author" minOccurs="0"/>
- *         <element name="tag" type="{http://www.w3.org/2001/XMLSchema}string" maxOccurs="unbounded" minOccurs="0"/>
- *       </sequence>
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "BlogPost", propOrder = {
-    "id",
-    "title",
-    "content",
-    "timestamp",
-    "author",
-    "tags"
-})
-public class BlogPost {
-
-    @XmlElement(required = true)
-    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-    @XmlID
-    @XmlSchemaType(name = "ID")
-    protected String id;
-    @XmlElement(required = true)
-    protected String title;
-    @XmlElement(required = true)
-    protected String content;
-    @XmlElement(required = true)
-    @XmlSchemaType(name = "dateTime")
-    protected XMLGregorianCalendar timestamp;
-    protected Author author;
-    @XmlElement(name = "tag")
-    protected List<String> tags;
-
-    /**
-     * Ruft den Wert der id-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getId() {
-        return id;
-    }
-
-    /**
-     * Legt den Wert der id-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setId(String value) {
-        this.id = value;
-    }
-
-    /**
-     * Ruft den Wert der title-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getTitle() {
-        return title;
-    }
-
-    /**
-     * Legt den Wert der title-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setTitle(String value) {
-        this.title = value;
-    }
-
-    /**
-     * Ruft den Wert der content-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link String }
-     *     
-     */
-    public String getContent() {
-        return content;
-    }
-
-    /**
-     * Legt den Wert der content-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link String }
-     *     
-     */
-    public void setContent(String value) {
-        this.content = value;
-    }
-
-    /**
-     * Ruft den Wert der timestamp-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public XMLGregorianCalendar getTimestamp() {
-        return timestamp;
-    }
-
-    /**
-     * Legt den Wert der timestamp-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link XMLGregorianCalendar }
-     *     
-     */
-    public void setTimestamp(XMLGregorianCalendar value) {
-        this.timestamp = value;
-    }
-
-    /**
-     * Ruft den Wert der author-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link Author }
-     *     
-     */
-    public Author getAuthor() {
-        return author;
-    }
-
-    /**
-     * Legt den Wert der author-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link Author }
-     *     
-     */
-    public void setAuthor(Author value) {
-        this.author = value;
-    }
-
-    /**
-     * Gets the value of the tags property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the Jakarta XML Binding object.
-     * This is why there is not a {@code set} method for the tags property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getTags().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link String }
-     * 
-     * 
-     * @return
-     *     The value of the tags property.
-     */
-    public List<String> getTags() {
-        if (tags == null) {
-            tags = new ArrayList<>();
-        }
-        return this.tags;
-    }
-
-}
