+++ service-provider/pom.xml
@@ -53,6 +53,11 @@
       <artifactId>spring-boot-starter-web-services</artifactId>
     </dependency>
     <dependency>
+      <groupId>de.samples.api-comparison</groupId>
+      <artifactId>api-stubs-soap</artifactId>
+      <version>0.0.1-SNAPSHOT</version>
+    </dependency>
+    <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-actuator</artifactId>
     </dependency>
@@ -125,26 +130,6 @@
   </dependencies>
 
   <build>
-    <pluginManagement>
-      <plugins>
-        <!--
-         to generate SOAP stubs, run mvn jaxws:wsimport
-         (if the plugin ignores the wsdl, delete /target/jaxws/stale
-         -->
-        <plugin>
-          <groupId>com.sun.xml.ws</groupId>
-          <artifactId>jaxws-maven-plugin</artifactId>
-          <version>4.0.2</version>
-          <configuration>
-            <wsdlDirectory>${project.basedir}/src/main/resources/soap/</wsdlDirectory>
-            <sourceDestDir>src/main/java</sourceDestDir>
-            <bindingFiles>
-              <bindingFile>${project.basedir}/src/main/resources/soap/wsimport.xjb</bindingFile>
-            </bindingFiles>
-          </configuration>
-        </plugin>
-      </plugins>
-    </pluginManagement>
     <plugins>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
