+++ service-provider/src/main/java/de/samples/apicomparison/provider/boundary/soap/mappers/AuthorSoapStubMapper.java
@@ -6,6 +6,6 @@ import org.mapstruct.Mapper;
 @Mapper(componentModel = "spring")
 public interface AuthorSoapStubMapper {
 
-  de.samples.apicomparison.provider.boundary.soap.stub.model.Author map(Author source);
+  de.samples.apicomparison.stubs.soap.model.Author map(Author source);
 
 }
