+++ service-provider/src/main/java/de/samples/apicomparison/provider/boundary/soap/mappers/BlogPostSoapStubMapper.java
@@ -1,7 +1,6 @@
 package de.samples.apicomparison.provider.boundary.soap.mappers;
 
 import de.samples.apicomparison.provider.domain.model.BlogPost;
-import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
@@ -14,9 +13,9 @@ import org.mapstruct.Mapping;
 )
 public interface BlogPostSoapStubMapper {
 
-  de.samples.apicomparison.provider.boundary.soap.stub.model.BlogPost map(BlogPost source);
+  de.samples.apicomparison.stubs.soap.model.BlogPost map(BlogPost source);
 
   @Mapping(target = "author", ignore = true)
-  BlogPost map(de.samples.apicomparison.provider.boundary.soap.stub.messages.BlogPostInput source);
+  BlogPost map(de.samples.apicomparison.stubs.soap.messages.BlogPostInput source);
 
 }
