--- service-provider/src/main/java/de/samples/apicomparison/provider/boundary/soap/stub/messages/CreateBlogPostRequest.java
@@ -1,64 +1,0 @@
-
-package de.samples.apicomparison.provider.boundary.soap.stub.messages;
-
-import jakarta.xml.bind.annotation.XmlAccessType;
-import jakarta.xml.bind.annotation.XmlAccessorType;
-import jakarta.xml.bind.annotation.XmlElement;
-import jakarta.xml.bind.annotation.XmlRootElement;
-import jakarta.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java-Klasse für anonymous complex type.
- * 
- * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
- * 
- * <pre>{@code
- * <complexType>
- *   <complexContent>
- *     <restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       <sequence>
- *         <element name="input" type="{http://samples.de/spring/soap/blog/messages}BlogPostInput"/>
- *       </sequence>
- *     </restriction>
- *   </complexContent>
- * </complexType>
- * }</pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-    "input"
-})
-@XmlRootElement(name = "createBlogPostRequest")
-public class CreateBlogPostRequest {
-
-    @XmlElement(required = true)
-    protected BlogPostInput input;
-
-    /**
-     * Ruft den Wert der input-Eigenschaft ab.
-     * 
-     * @return
-     *     possible object is
-     *     {@link BlogPostInput }
-     *     
-     */
-    public BlogPostInput getInput() {
-        return input;
-    }
-
-    /**
-     * Legt den Wert der input-Eigenschaft fest.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link BlogPostInput }
-     *     
-     */
-    public void setInput(BlogPostInput value) {
-        this.input = value;
-    }
-
-}
