--- service-provider/src/main/java/de/samples/apicomparison/provider/boundary/soap/stub/messages/ObjectFactory.java
@@ -1,122 +1,0 @@
-
-package de.samples.apicomparison.provider.boundary.soap.stub.messages;
-
-import jakarta.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the de.samples.apicomparison.provider.boundary.soap.stub.messages package. 
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: de.samples.apicomparison.provider.boundary.soap.stub.messages
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link DeleteBlogPostResponse }
-     * 
-     * @return
-     *     the new instance of {@link DeleteBlogPostResponse }
-     */
-    public DeleteBlogPostResponse createDeleteBlogPostResponse() {
-        return new DeleteBlogPostResponse();
-    }
-
-    /**
-     * Create an instance of {@link FindAllBlogPostsRequest }
-     * 
-     * @return
-     *     the new instance of {@link FindAllBlogPostsRequest }
-     */
-    public FindAllBlogPostsRequest createFindAllBlogPostsRequest() {
-        return new FindAllBlogPostsRequest();
-    }
-
-    /**
-     * Create an instance of {@link FindAllBlogPostsResponse }
-     * 
-     * @return
-     *     the new instance of {@link FindAllBlogPostsResponse }
-     */
-    public FindAllBlogPostsResponse createFindAllBlogPostsResponse() {
-        return new FindAllBlogPostsResponse();
-    }
-
-    /**
-     * Create an instance of {@link CreateBlogPostRequest }
-     * 
-     * @return
-     *     the new instance of {@link CreateBlogPostRequest }
-     */
-    public CreateBlogPostRequest createCreateBlogPostRequest() {
-        return new CreateBlogPostRequest();
-    }
-
-    /**
-     * Create an instance of {@link BlogPostInput }
-     * 
-     * @return
-     *     the new instance of {@link BlogPostInput }
-     */
-    public BlogPostInput createBlogPostInput() {
-        return new BlogPostInput();
-    }
-
-    /**
-     * Create an instance of {@link CreateBlogPostResponse }
-     * 
-     * @return
-     *     the new instance of {@link CreateBlogPostResponse }
-     */
-    public CreateBlogPostResponse createCreateBlogPostResponse() {
-        return new CreateBlogPostResponse();
-    }
-
-    /**
-     * Create an instance of {@link DeleteBlogPostRequest }
-     * 
-     * @return
-     *     the new instance of {@link DeleteBlogPostRequest }
-     */
-    public DeleteBlogPostRequest createDeleteBlogPostRequest() {
-        return new DeleteBlogPostRequest();
-    }
-
-    /**
-     * Create an instance of {@link FindBlogPostByIdRequest }
-     * 
-     * @return
-     *     the new instance of {@link FindBlogPostByIdRequest }
-     */
-    public FindBlogPostByIdRequest createFindBlogPostByIdRequest() {
-        return new FindBlogPostByIdRequest();
-    }
-
-    /**
-     * Create an instance of {@link FindBlogPostByIdResponse }
-     * 
-     * @return
-     *     the new instance of {@link FindBlogPostByIdResponse }
-     */
-    public FindBlogPostByIdResponse createFindBlogPostByIdResponse() {
-        return new FindBlogPostByIdResponse();
-    }
-
-}
