--- service-provider/src/main/java/de/samples/apicomparison/provider/boundary/soap/stub/service/BlogPostServiceInterface.java
@@ -1,83 +1,0 @@
-
-package de.samples.apicomparison.provider.boundary.soap.stub.service;
-
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.CreateBlogPostRequest;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.CreateBlogPostResponse;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.DeleteBlogPostRequest;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.DeleteBlogPostResponse;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.FindAllBlogPostsRequest;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.FindAllBlogPostsResponse;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.FindBlogPostByIdRequest;
-import de.samples.apicomparison.provider.boundary.soap.stub.messages.FindBlogPostByIdResponse;
-import jakarta.jws.WebMethod;
-import jakarta.jws.WebParam;
-import jakarta.jws.WebResult;
-import jakarta.jws.WebService;
-import jakarta.jws.soap.SOAPBinding;
-import jakarta.xml.bind.annotation.XmlSeeAlso;
-
-
-/**
- * This class was generated by the XML-WS Tools.
- * XML-WS Tools 4.0.1
- * Generated source version: 3.0
- * 
- */
-@WebService(name = "BlogPostServiceInterface", targetNamespace = "http://samples.de/spring/soap/blog/service")
-@SOAPBinding(parameterStyle = SOAPBinding.ParameterStyle.BARE)
-@XmlSeeAlso({
-    de.samples.apicomparison.provider.boundary.soap.stub.messages.ObjectFactory.class,
-    de.samples.apicomparison.provider.boundary.soap.stub.model.ObjectFactory.class
-})
-public interface BlogPostServiceInterface {
-
-
-    /**
-     * 
-     * @param findAllInputPart
-     * @return
-     *     returns de.samples.apicomparison.provider.boundary.soap.stub.messages.FindAllBlogPostsResponse
-     */
-    @WebMethod(action = "findAll")
-    @WebResult(name = "findAllBlogPostsResponse", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "findAllOutputPart")
-    public FindAllBlogPostsResponse findAll(
-        @WebParam(name = "findAllBlogPostsRequest", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "findAllInputPart")
-        FindAllBlogPostsRequest findAllInputPart);
-
-    /**
-     * 
-     * @param createInputPart
-     * @return
-     *     returns de.samples.apicomparison.provider.boundary.soap.stub.messages.CreateBlogPostResponse
-     */
-    @WebMethod(action = "create")
-    @WebResult(name = "createBlogPostResponse", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "createOutputPart")
-    public CreateBlogPostResponse create(
-        @WebParam(name = "createBlogPostRequest", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "createInputPart")
-        CreateBlogPostRequest createInputPart);
-
-    /**
-     * 
-     * @param deleteInputPart
-     * @return
-     *     returns de.samples.apicomparison.provider.boundary.soap.stub.messages.DeleteBlogPostResponse
-     */
-    @WebMethod(action = "deleteById")
-    @WebResult(name = "deleteBlogPostResponse", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "deleteOutputPart")
-    public DeleteBlogPostResponse deleteById(
-        @WebParam(name = "deleteBlogPostRequest", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "deleteInputPart")
-        DeleteBlogPostRequest deleteInputPart);
-
-    /**
-     * 
-     * @param findByIdInputPart
-     * @return
-     *     returns de.samples.apicomparison.provider.boundary.soap.stub.messages.FindBlogPostByIdResponse
-     */
-    @WebMethod(action = "findById")
-    @WebResult(name = "findBlogPostByIdResponse", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "findByIdOutputPart")
-    public FindBlogPostByIdResponse findById(
-        @WebParam(name = "findBlogPostByIdRequest", targetNamespace = "http://samples.de/spring/soap/blog/messages", partName = "findByIdInputPart")
-        FindBlogPostByIdRequest findByIdInputPart);
-
-}
