--- service-provider/src/main/resources/soap/wsimport.xjb
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jaxws:bindings
-  xmlns:jaxws="https://jakarta.ee/xml/ns/jaxws"
-  xmlns:jaxb="https://jakarta.ee/xml/ns/jaxb"
-  xmlns:xjc="http://java.sun.com/xml/ns/jaxb/xjc"
-  xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/"
-  xmlns:xs="http://www.w3.org/2001/XMLSchema" >
-  <jaxws:package name="de.samples.apicomparison.provider.boundary.soap.stub.service"/>
-  <jaxb:globalBindings>
-    <xjc:simple/>
-  </jaxb:globalBindings>
-  <jaxws:bindings
-    node="wsdl:definitions/wsdl:types/xs:schema[@targetNamespace='http://samples.de/spring/soap/blog/messages']">
-    <jaxb:globalBindings>
-      <xjc:simple/>
-    </jaxb:globalBindings>
-    <jaxb:schemaBindings>
-      <jaxb:package name="de.samples.apicomparison.provider.boundary.soap.stub.messages"/>
-    </jaxb:schemaBindings>
-    <jaxb:bindings node="//xs:element[@name='findAllBlogPostsResponse']/xs:complexType/xs:sequence/xs:element[@name='blogPost']">
-      <jaxb:property name="blogPosts"/>
-    </jaxb:bindings>
-    <jaxb:bindings node="//xs:element[@name='deleteBlogPostResponse']/xs:complexType/xs:sequence/xs:element[@name='status']/xs:simpleType">
-      <jaxb:typesafeEnumClass name="DeleteOperationStatus"/>
-    </jaxb:bindings>
-  </jaxws:bindings>
-  <jaxws:bindings
-    node="wsdl:definitions/wsdl:types/xs:schema[@targetNamespace='http://samples.de/spring/soap/blog/types']">
-    <jaxb:globalBindings>
-      <xjc:simple/>
-    </jaxb:globalBindings>
-    <jaxb:schemaBindings>
-      <jaxb:package name="de.samples.apicomparison.provider.boundary.soap.stub.model"/>
-    </jaxb:schemaBindings>
-  </jaxws:bindings>
-</jaxws:bindings>
